requirements.txt
beautifulsoup4==4.14.2
cloudscraper==1.2.71
cursor==1.3.5
google_api_python_client==2.179.0
google_auth_oauthlib==1.2.2
google-cloud-texttospeech==2.16.3
moviepy==1.0.3
numpy<2
opencv_python==4.7.0.72
opencv_python_headless==4.8.0.74
Pillow==12.0.0
proglog==0.1.12
protobuf>=3.19.5,<5.0.0
psutil==7.0.0
Requests
wurlitzer==3.0.3
edge-tts
Requests[socks]

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0